smokebot: minor edit

+++ Utilities/Smokebot/run_smokebot.sh
@@ -191,9 +191,9 @@ if [ "$KILL_SMOKEBOT" == "1" ]; then
     kill -9 $PID
     ../../Verification/scripts/Run_SMV_Cases.sh -s  >& /dev/null
     echo smokebot process $PID killed
-    echo before remove
-    rm $smokebot_pid
-    echo after remove
+    if [ -e $smokebot_pid ]; then
+      rm $smokebot_pid
+    fi
   else
     echo smokebotbot is not running, cannot be killed.
   fi
